Update docs, restructure repo
