+++ gym_roboy/envs/__init__.py
@@ -1,2 +1,1 @@
-from .simulation_client import SimulationClient, RosSimulationClient, StubSimulationClient
 from .roboy_env import RoboyEnv
