+++ gym_roboy/envs/tests/run_all_tests.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+python3 -m pytest --run-integration -v -x $@  # -x to stop on the first failed test
