--- gym_roboy/envs/tests/run_tests.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-echo "Please source ROS2 yourself"
-python3 -m pytest --run-integration --disable-warnings -v -x $@
